+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/EnvelopeNeoForge.java
@@ -1,14 +1,20 @@
 package io.github.mortuusars.envelope.neoforge;
 
 import com.google.common.base.Preconditions;
+import com.mojang.serialization.MapCodec;
 import io.github.mortuusars.envelope.Config;
 import io.github.mortuusars.envelope.Envelope;
+import io.github.mortuusars.envelope.neoforge.loot.ConfigurableAddTableLootModifier;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.config.ModConfig;
 import net.neoforged.fml.loading.FMLEnvironment;
+import net.neoforged.neoforge.common.loot.IGlobalLootModifier;
+import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.neoforge.registries.DeferredRegister;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
 import org.jetbrains.annotations.Nullable;
 
 @Mod(Envelope.ID)
@@ -17,6 +23,7 @@ public class EnvelopeNeoForge {
         Envelope.init();
 
         container.registerConfig(ModConfig.Type.SERVER, Config.Server.SPEC);
+        container.registerConfig(ModConfig.Type.COMMON, Config.Common.SPEC);
         container.registerConfig(ModConfig.Type.CLIENT, Config.Client.SPEC);
 
         @Nullable IEventBus modEventBus = container.getEventBus();
@@ -39,9 +46,18 @@ public class EnvelopeNeoForge {
         RegisterImpl.DATA_COMPONENT_TYPES.register(modEventBus);
         RegisterImpl.PARTICLE_TYPES.register(modEventBus);
         RegisterImpl.CUSTOM_STATS.register(modEventBus);
+        LootModifiers.LOOT_MODIFIERS.register(modEventBus);
 
         if (FMLEnvironment.dist == Dist.CLIENT) {
             EnvelopeNeoForgeClient.init(container);
         }
+    }
+
+    public static class LootModifiers {
+        private static final DeferredRegister<MapCodec<? extends IGlobalLootModifier>> LOOT_MODIFIERS =
+              DeferredRegister.create(NeoForgeRegistries.Keys.GLOBAL_LOOT_MODIFIER_SERIALIZERS, Envelope.ID);
+
+        public static final DeferredHolder<MapCodec<? extends IGlobalLootModifier>, MapCodec<ConfigurableAddTableLootModifier>> ADD_TABLE =
+              LOOT_MODIFIERS.register("add_table", () -> ConfigurableAddTableLootModifier.CODEC);
     }
 }