+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/datagen/server/loot/ChestLootProvider.java
@@ -10,7 +10,10 @@ import net.minecraft.core.component.DataComponents;
 import net.minecraft.data.loot.LootTableSubProvider;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.Style;
 import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.Unit;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.component.SeededContainerLoot;
 import net.minecraft.world.level.block.Blocks;
@@ -21,6 +24,7 @@ import net.minecraft.world.level.storage.loot.entries.LootItem;
 import net.minecraft.world.level.storage.loot.entries.LootPoolSingletonContainer;
 import net.minecraft.world.level.storage.loot.entries.NestedLootTable;
 import net.minecraft.world.level.storage.loot.functions.EnchantWithLevelsFunction;
+import net.minecraft.world.level.storage.loot.functions.LootItemConditionalFunction;
 import net.minecraft.world.level.storage.loot.functions.SetComponentsFunction;
 import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
 import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
@@ -39,6 +43,11 @@ public class ChestLootProvider implements LootTableSubProvider {
 
     @Override
     public void generate(@NotNull BiConsumer<ResourceKey<LootTable>, LootTable.Builder> output) {
+        collapsedMailHub(output);
+        injects(output);
+    }
+
+    private void collapsedMailHub(@NotNull BiConsumer<ResourceKey<LootTable>, LootTable.Builder> output) {
         output.accept(
               Envelope.LootTables.COLLAPSED_MAIL_HUB_STORAGE,
               LootTable.lootTable().withPool(LootPool.lootPool()
@@ -189,9 +198,40 @@ public class ChestLootProvider implements LootTableSubProvider {
         );
     }
 
+    private void injects(@NotNull BiConsumer<ResourceKey<LootTable>, LootTable.Builder> output) {
+        output.accept(
+              Envelope.LootTables.ABANDONED_MINESHAFT_INJECT,
+              LootTable.lootTable()
+                    .withPool(LootPool.lootPool()
+                          .setRolls(ConstantValue.exactly(1))
+                          .add(LootItem.lootTableItem(Envelope.Items.PAPER_BOX.get()).apply(SetItemCountFunction.setCount(UniformGenerator.between(1.0F, 3.0F))))
+                          .add(EmptyLootItem.emptyItem().setWeight(2)))
+                    .withPool(LootPool.lootPool()
+                          .add(letter(Component.translatable("letter.envelope.abandoned_mineshaft.journal_page.name"), Component.translatable("letter.envelope.abandoned_mineshaft.journal_page_1")).apply(tattered()))
+                          .add(letter(Component.translatable("letter.envelope.abandoned_mineshaft.journal_page.name"), Component.translatable("letter.envelope.abandoned_mineshaft.journal_page_2")).apply(tattered()))
+                          .add(letter(Component.translatable("letter.envelope.abandoned_mineshaft.journal_page.name"), Component.translatable("letter.envelope.abandoned_mineshaft.journal_page_3")).apply(tattered()))
+                          .add(EmptyLootItem.emptyItem().setWeight(16))
+                    ));
+
+        output.accept(
+              Envelope.LootTables.PILLAGER_OUTPOST_INJECT,
+              LootTable.lootTable()
+                    .withPool(LootPool.lootPool()
+                          .setRolls(ConstantValue.exactly(1))
+                          .add(LootItem.lootTableItem(Envelope.Items.PAPER_BOX.get()).apply(SetItemCountFunction.setCount(UniformGenerator.between(1.0F, 3.0F))))
+                          .add(EmptyLootItem.emptyItem().setWeight(2)))
+                    .withPool(LootPool.lootPool()
+                          .add(illagerLetter(Component.translatable("letter.envelope.pillager_outpost.report.name"), Component.translatable("letter.envelope.pillager_outpost.report_1")).apply(tattered()))
+                          .add(illagerLetter(Component.translatable("letter.envelope.pillager_outpost.report.name"), Component.translatable("letter.envelope.pillager_outpost.report_2")).apply(tattered()))
+                          .add(illagerLetter(Component.translatable("letter.envelope.pillager_outpost.report.name"), Component.translatable("letter.envelope.pillager_outpost.report_3")).apply(tattered()))
+                          .add(illagerLetter(Component.translatable("letter.envelope.pillager_outpost.report.name"), Component.translatable("letter.envelope.pillager_outpost.report_4")).apply(tattered()))
+                          .add(EmptyLootItem.emptyItem().setWeight(16))
+                    ));
+    }
+
     // --
 
-    protected LootPoolSingletonContainer.Builder<?> letter(@Nullable Component name, @NotNull Component text) {
+    protected LootPoolSingletonContainer.Builder<?> letter(@Nullable MutableComponent name, @NotNull MutableComponent text) {
         LootPoolSingletonContainer.Builder<?> builder = LootItem.lootTableItem(Envelope.Items.LETTER.get())
               .apply(SetComponentsFunction.setComponent(Envelope.DataComponents.LETTER_CONTENT, new LetterContent(text)));
 
@@ -200,5 +240,22 @@ public class ChestLootProvider implements LootTableSubProvider {
         }
 
         return builder;
+    }
+
+    protected LootPoolSingletonContainer.Builder<?> illagerLetter(@Nullable MutableComponent name, @NotNull MutableComponent text) {
+        LootPoolSingletonContainer.Builder<?> builder = LootItem.lootTableItem(Envelope.Items.LETTER.get())
+              .apply(SetComponentsFunction.setComponent(Envelope.DataComponents.LETTER_CONTENT, new LetterContent(text
+                    .withStyle(Style.EMPTY.withFont(ResourceLocation.withDefaultNamespace("illageralt"))))));
+
+        if (name != null) {
+            builder.apply(SetComponentsFunction.setComponent(DataComponents.ITEM_NAME, name
+                  .withStyle(Style.EMPTY.withFont(ResourceLocation.withDefaultNamespace("illageralt")))));
+        }
+
+        return builder;
+    }
+
+    private static LootItemConditionalFunction.@NotNull Builder<?> tattered() {
+        return SetComponentsFunction.setComponent(Envelope.DataComponents.LETTER_TATTERED, Unit.INSTANCE);
     }
 }